% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwaynames-package.R
\docType{data}
\name{all}
\alias{all}
\title{All Norwegian names.}
\format{A data frame with five variables: \code{first_name}, \code{year},
  \code{Percent} (the proportion of births in that year with that name),
  \code{Count} (the number of births in that year with that name)and \code{Sex}.}
\usage{
all
}
\description{
Full baby name data provided by Statistics Norway. This includes all names
with at least 4 uses.
}
\keyword{datasets}
