% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norwaynames-package.R
\docType{package}
\name{norwaynames-package}
\alias{norwaynames}
\alias{norwaynames-package}
\title{norwaynames: Provides Norweigian Baby Name Data}
\description{
this dataset includes the sex, count, and proportion of births for 
              each first name from 1880 - 2017. The data set includes all names 
              with more than 4 instances in any given year.
}
\author{
\strong{Maintainer}: Emily Kothe \email{ekothe@deakin.edu.au}

}
\keyword{internal}
